refactor(jsoneditor): big胆 refactor jsoneditor.py
