+++ BAAH_CONFIGS/example.json
@@ -262,5 +262,7 @@
     "SPEICAL_EVENT_STATUS": false,
     "NORMAL_QUEST_EVENT_STATUS": false,
     "HARD_QUEST_EVENT_STATUS": false,
-    "EXCHANGE_EVENT_STATUS": false
+    "EXCHANGE_EVENT_STATUS": false,
+    "CONTEST_LEVEL_PRIORITY": 10,
+    "CONTEST_RANK_PRIORITY": 10
 }
