+++ gui/__init__.py
@@ -1,4 +1,5 @@
 from .show_gui import show_gui
+from .refactor_pages.json_file_docker import tab_area
 
 
-__all__ = ["show_gui"]
+__all__ = ["show_gui", "tab_area"]
