+++ gui/define.py
@@ -3,6 +3,7 @@ from common import TaskName
 
 # 构造一个config，用于在tab间共享softwareconfigdict
 gui_shared_config = MyConfigger()
+curr_config = MyConfigger()
 
 
 class TaskStr:
